manager updated
